# example-microservice/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Delivery Specification
  description: This is a microservice for managing delivery in the YumYumNow company.
  version: 1.0.0
servers:
  - url: http://localhost:8080
...
paths:
...
  /admin/exceptions:
    get:
      summary: Retrieve all exceptions
      operationId: getExceptions
      tags:
        - Admin
      description: Return a list of all exceptions collected from orders. To be used by admin.
      parameters:
        - name: authorization
          in: query
          description: the userId to check if they have the rights to make this request
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful response, all exceptions received
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeliveryException'
        '400':
          description: Unsuccessful, exceptions cannot be retrieved because of a bad request
        '403':
          description: Unsuccessful, entity does not have access rights to retrieve exceptions
        '404':
          description: Unsuccessful, no exceptions were found
  /admin/exceptions/{orderId}:
    get:
      summary: Retrieve exception for a specific order
      operationId: getExceptionForOrder
      tags:
        - Admin
      description: Return exception information for a specific order. To be used by admin.
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: authorization
          in: query
          description: the userId to check if they have the rights to make this request
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful response, exception for the specific order received
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeliveryException'
        '400':
          description: Unsuccessful, specific exception cannot be retrieved because of a bad request
        '403':
          description: Unsuccessful, entity does not have access rights to retrieve specific exception
        '404':
          description: Unsuccessful, no specific exception was found
    post:
      summary: Add exception for a specific order
      operationId: makeException
      tags:
        - Admin
      description: Add exception information for a specific order. To be used by admin.
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: authorization
          in: query
          description: The userId to check if they have the rights to make this request
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeliveryException'
      responses:
        '200':
          description: Successful response, exception for the specific order added
        '400':
          description: Unsuccessful, specific exception cannot be added because of a bad request
        '403':
          description: Unsuccessful, entity does not have access rights to add specific exception
        '404':
          description: Unsuccessful, no specific exception was found
    put:
      summary: Update exception for a specific order
      operationId: updateException
      tags:
        - Admin
      description: Update exception information for a specific order. To be used by admin.
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: authorization
          in: query
          description: The userId to check if they have the rights to make this request
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeliveryException'
      responses:
        '200':
          description: Successful response, exception for the specific order updated
        '400':
          description: Unsuccessful, specific exception cannot be updated because of a bad request
        '403':
          description: Unsuccessful, entity does not have access rights to update specific exception
        '404':
          description: Unsuccessful, no specific exception was found
  /admin/analytics/courier-efficiency:
    get:
      summary: Retrieve the courier efficiencies of all couriers
      operationId: getCourierEfficiencies
      tags:
        - Admin
      description: Return the all courier efficiencies.
      parameters:
        - name: authorization
          in: query
          description: the userId to check if they have the rights to make this request, only admin is allowed.
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful response, driver efficiencies received
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: number
                  format: double
                  description: the number represents a percentage of a particular driver's efficiency
        '400':
          description: Unsuccessful, driver efficiencies cannot be retrieved because of a bad request
        '403':
          description: Unsuccessful, entity does not have access rights to retrieve driver efficiencies
        '404':
          description: Unsuccessful, driver efficiencies were not found
  /admin/analytics/all-ratings:
    get:
      summary: Retrieve a list of all ratings
      operationId: getAllRatings
      tags:
        - Admin
      description: Return rating of all of the orders.
      parameters:
        - name: authorization
          in: query
          description: the userId to check if they have the rights to make this request, only admin is allowed.
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful response, ratings received
          content:
            application/json:
              schema:
                type: array
                items:
                  type: number
                  format: Long
        '400':
          description: Unsuccessful, ratings cannot be retrieved because of a bad request
        '403':
          description: Unsuccessful, entity does not have access rights to retrieve ratings
        '404':
          description: Unsuccessful, ratings were not found
...
  /admin/vendor/radius:
    get:
      summary: Retrieve the default radius for vendors that don't have their own couriers (independent vendors)
      operationId: getDefaultRadius
      tags:
        - Admin
      description: Return the default radius of all independent vendors. To be used by admin.
      parameters:
        - name: authorization
          in: query
          description: the userId to check if they have the rights to make this request
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful response, default radius of vendors with independent couriers received
          content:
            application/json:
              schema:
                type: number
                format: double
        '400':
          description: Unsuccessful, default radius cannot be retrieved because of a bad request
        '403':
          description: Unsuccessful, entity does not have access rights to retrieve specific radius
        '404':
          description: Unsuccessful, no specific radius was found
    put:
      summary: Update the default radius for vendors that don't have their own couriers (independent vendors)
      operationId: updateDefaultRadius
      tags:
        - Admin
      description: Update the default radius of all independent vendors. To be used by admin.
      parameters:
        - name: authorization
          in: query
          description: The userId to check if they have the rights to make this request
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Radius in meters to be set as a new value
        required: true
        content:
          application/json:
            schema:
              type: number
              format: double
      responses:
        '200':
          description: Successful response, default radius of vendors with independent couriers updated
        '400':
          description: Unsuccessful, default radius cannot be updated because of a bad request
        '403':
          description: Unsuccessful, entity does not have access rights to update radius. Only admins have this right.
        '404':
          description: Unsuccessful, no default radius was found
...
components:
  schemas:
    Order:
      type: object
      x-class-extra-annotation: "@javax.persistence.Entity @javax.persistence.Table(name = \"order_table\") @lombok.AllArgsConstructor @lombok.NoArgsConstructor"
      x-field-extra-annotation: "@javax.persistence.OneToOne(cascade = javax.persistence.CascadeType.ALL) @javax.persistence.JoinColumn(name =\"orderId\")"
      properties:
        status:
          type: string
          enum:
            - pending
            - accepted
            - rejected
            - preparing
            - given-to-courier
            - in-transit
            - delivered
          description: the current status of the order expressed as one of the above options
        id:
          type: integer
          format: int64
          description: the unique id of an order
          x-field-extra-annotation: "@javax.persistence.Id"
        courierId:
          type: integer
          format: int64
          example: 88
          description: the id of the courier that is delivering the order
        vendorId:
          type: integer
          format: int64
          example: 99
          description: the id of the vendor that prepares the order
        deliveryDestination:
          $ref: '#/components/schemas/Location'
        timeValues:
          $ref: '#/components/schemas/Time'
        ratingNumber:
          type: number
          format: Long
          description: a number out of five (0-5 stars)
...
    DeliveryException:
      description: >-
        General exception class more specific exceptions inherit from.
        Thrown whenever there is any kind of problem with the order
      x-class-extra-annotation: "@javax.persistence.Entity @javax.persistence.Table(name= \"exception_table\") @lombok.AllArgsConstructor @lombok.NoArgsConstructor"
      properties:
        id:
          type: integer
          format: int64
          x-field-extra-annotation: "@javax.persistence.Id @javax.persistence.GeneratedValue(strategy = javax.persistence.GenerationType.IDENTITY)"
        order:
          $ref: '#/components/schemas/Order'
        message:
          type: string
          description: explanation of what went wrong with the order
          example: "Order was delivered laaaate"
        isResolved:
          type: boolean
          example: true
        exceptionType:
          type: string
          enum:
            - LateDelivery
            - NotDelivered
            - Rejected
            - Other
          description: The kind of unforeseen situation of the order
          example: LateDelivery
    Location:
      description: Any kind of address expressed as a longitude and latitude
      type: object
      x-class-extra-annotation: "@javax.persistence.Embeddable"
      x-field-extra-annotation: "@javax.persistence.Embedded"
      properties:
        latitude:
          type: number
          format: float
          example: 23.92488
        longitude:
          type: number
          format: float
          example: 74.7099
    Time:
      description: Collection of time attributes of an order
      type: object
      x-class-extra-annotation: "@javax.persistence.Embeddable"
      x-field-extra-annotation: "@javax.persistence.Embedded"
      properties:
        prepTime:
          type: string
          format: time
          description: the time it takes to prepare the order in format "HH:MM:SS"
          example: '01:30:00'
        expectedDeliveryTime:
          type: string
          format: date-time
          description: the time the order will be planned to be delivered
        actualDeliveryTime:
          type: string
          format: date-time
          description: the time the order actually is delivered
        orderTime:
          type: string
          format: date-time
          description: the time the order was placed
...